Change chance to double

+++ OTMonsterConverter/MonsterTypes/Summon.cs
@@ -8,7 +8,7 @@ namespace OTMonsterConverter.MonsterTypes
     {
         public string Name { get; set; }
         public uint Rate { get; set; }
-        public uint Chance { get; set; }
+        public double Chance { get; set; }
         public int Max { get; set; }
         public bool Force { get; set; }
     }
